Refactored providers so they are now in dedicated .c files for easier maintenance Added -P/--providers to man page
