+++ src/pvdr/kprobe.c
@@ -39,31 +39,27 @@
 #include <ply/ply.h>
 #include <ply/pvdr.h>
 
-#define KPROBE_MAXLEN   0x100
+#include "kprobe.h"
 
-/*
- * Structure used for internal representation of kprobes, kretprobes and
- * tracepoints.
- */
-typedef struct kprobe {
-	const char *pvdr;
-	const char *type;
-	FILE *ctrl;
-	int bfd;
+const module_t *kprobe_modules[] = {
+	&kprobe_module,
 
-	struct {
-		int cap, len;
-		int *fds;
-	} efds;
-} kprobe_t;
+	&method_module,
+	&common_module,
 
-typedef struct profile {
-	int *efds;
-	int num;
-	kprobe_t *kp;
-} profile_t;
+	NULL
+};
 
-static int probe_event_id(kprobe_t *kp, const char *path)
+const module_t *kretprobe_modules[] = {
+	&kretprobe_module,
+
+	&method_module,
+	&common_module,
+
+	NULL
+};
+
+int probe_event_id(kprobe_t *kp, const char *path)
 {
 	FILE *fp;
 	char ev_id[16];
@@ -79,7 +75,7 @@ static int probe_event_id(kprobe_t *kp, const char *path)
 	return strtol(ev_id, NULL, 0);
 }
 
-static int probe_attach(kprobe_t *kp, int id)
+int probe_attach(kprobe_t *kp, int id)
 {
 	struct perf_event_attr attr = {};
 	int efd, gfd;
@@ -122,7 +118,7 @@ static int probe_attach(kprobe_t *kp, int id)
 	return 1;
 }
 
-static kprobe_t *probe_load(enum bpf_prog_type type,
+kprobe_t *probe_load(enum bpf_prog_type type,
 			    node_t *probe, prog_t *prog)
 {
 	kprobe_t *kp;
@@ -153,7 +149,7 @@ static kprobe_t *probe_load(enum bpf_prog_type type,
 	return kp;
 }
 
-static int probe_teardown_events(kprobe_t *kp)
+int probe_teardown_events(kprobe_t *kp)
 {
 	int i;
 
@@ -164,7 +160,7 @@ static int probe_teardown_events(kprobe_t *kp)
 	return 0;
 }
 
-static int probe_teardown(node_t *probe)
+int probe_teardown(node_t *probe)
 {
 	int err;
 
@@ -198,7 +194,7 @@ static int probe_teardown(node_t *probe)
  * active for different instances of ply running simultaneously, and it
  * gives us a way to clean up after ourselves only when done.
  */
-static int kprobe_setattach(kprobe_t *kp, const char *func_and_offset,
+int kprobe_setattach(kprobe_t *kp, const char *func_and_offset,
 			    int attach)
 {
 	char fullprobename[KPROBE_MAXLEN];
@@ -290,7 +286,7 @@ static int kprobe_setattach(kprobe_t *kp, const char *func_and_offset,
 	return probe_attach(kp, id);
 }
 
-static int kprobe_setattach_pattern(kprobe_t *kp, const char *pattern,
+int kprobe_setattach_pattern(kprobe_t *kp, const char *pattern,
 				    int attach)
 {
 	int i, err;
@@ -321,7 +317,7 @@ static int kprobe_setattach_pattern(kprobe_t *kp, const char *pattern,
 	return (err < 0) ? err : kp->efds.len;
 }
 
-static int kprobe_load(node_t *probe, prog_t *prog, const char *probestring,
+int kprobe_load(node_t *probe, prog_t *prog, const char *probestring,
 		       const char *type, kprobe_t **kpp)
 {
 	kprobe_t *kp;
@@ -344,7 +340,7 @@ static int kprobe_load(node_t *probe, prog_t *prog, const char *probestring,
 	return kprobe_setattach_pattern(kp, func, 1);
 }
 
-static int kprobe_detach(kprobe_t *kp, const char *probestring)
+int kprobe_detach(kprobe_t *kp, const char *probestring)
 {
 	char *func;
 
@@ -352,15 +348,6 @@ static int kprobe_detach(kprobe_t *kp, const char *probestring)
 
 	return kprobe_setattach_pattern(kp, func, 0);
 }
-
-const module_t *kprobe_modules[] = {
-	&kprobe_module,
-
-	&method_module,
-	&common_module,
-
-	NULL
-};
 
 int kprobe_default(node_t *probe, node_t **stmts)
 {
@@ -389,18 +376,18 @@ int kprobe_default(node_t *probe, node_t **stmts)
 	return 0;
 }
 
-static int kprobe_resolve(node_t *call, const func_t **f)
+int kprobe_resolve(node_t *call, const func_t **f)
 {
 	return modules_get_func(kprobe_modules, call, f);
 }
 
-static int kprobe_setup(node_t *probe, prog_t *prog)
+int kprobe_setup(node_t *probe, prog_t *prog)
 {
 	return kprobe_load(probe, prog, probe->string, "p",
 			   (kprobe_t **)&probe->dyn->probe.pvdr_priv);
 }
 
-static int kprobe_destroy(kprobe_t *kp, const char *pattern)
+int kprobe_destroy(kprobe_t *kp, const char *pattern)
 {
 	int err1, err2;
 
@@ -421,7 +408,7 @@ static int kprobe_destroy(kprobe_t *kp, const char *pattern)
 	return err1 ? err1 : err2;
 }
 
-static int kprobe_teardown(node_t *probe)
+int kprobe_teardown(node_t *probe)
 {
 	kprobe_t *kp = probe->dyn->probe.pvdr_priv;
 
@@ -437,16 +424,6 @@ pvdr_t kprobe_pvdr = {
 
 	.setup      = kprobe_setup,
 	.teardown   = kprobe_teardown,
-};
-
-
-const module_t *kretprobe_modules[] = {
-	&kretprobe_module,
-
-	&method_module,
-	&common_module,
-
-	NULL
 };
 
 int kretprobe_default(node_t *probe, node_t **stmts)
@@ -476,12 +453,12 @@ int kretprobe_default(node_t *probe, node_t **stmts)
 	return 0;
 }
 
-static int kretprobe_resolve(node_t *call, const func_t **f)
+int kretprobe_resolve(node_t *call, const func_t **f)
 {
 	return modules_get_func(kretprobe_modules, call, f);
 }
 
-static int kretprobe_setup(node_t *probe, prog_t *prog)
+int kretprobe_setup(node_t *probe, prog_t *prog)
 {
 	return kprobe_load(probe, prog, probe->string, "r",
 			   (kprobe_t **)&probe->dyn->probe.pvdr_priv);
@@ -498,198 +475,10 @@ pvdr_t kretprobe_pvdr = {
 	.teardown   = kprobe_teardown,
 };
 
-/* PROFILE provider */
-
-static int profile_resolve(node_t *call, const func_t **f)
-{
-        return modules_get_func(kprobe_modules, call, f);
-}
-
-static void profile_destroy(profile_t *profile)
-{
-	int i;
-
-	if (!profile)
-		return;
-
-	if (profile->kp)
-		kprobe_destroy(profile->kp, "kprobe:perf_swevent_hrtimer");
-	for (i = 0; i < profile->num; i++) {
-		if (profile->efds[i] > 0)
-			close(profile->efds[i]);
-	}
-	free(profile->efds);
-	free(profile);
-}
-
-static int profile_teardown(node_t *probe)
-{
-	profile_t *profile = probe->dyn->probe.pvdr_priv;
-
-	profile_destroy(profile);
-
-	return 0;
-}
-
-static int profile_perf_event_open(profile_t *profile, int cpu, int freq)
-{
-	struct perf_event_attr attr = {};
-	int err = 0, i = profile->num;
-
-	attr.type = PERF_TYPE_SOFTWARE;
-	attr.config = PERF_COUNT_SW_CPU_CLOCK;
-	attr.freq = 1;
-	attr.sample_freq = freq;
-
-	profile->efds[i] = perf_event_open(&attr, -1, cpu, -1, 0);
-	if (profile->efds[i] < 0)
-		return -errno;
-	if (ioctl(profile->efds[i], PERF_EVENT_IOC_ENABLE, 0)) {
-		close(profile->efds[i]);
-		return -errno;
-	}
-	profile->num++;
-	return 0;
-}
-
-/*
- * profile provider is implemented by creating a perf event
- * PERF_TYPE_SOFTWARE/PERF_COUNT_SW_CPU_CLOCK for each CPU (or
- * a specified CPU) and using a kprobe on "kprobe:perf_swevent_hrtimer"
- * to catch it occuring in kernel context.
- */
-static int profile_setup(node_t *probe, prog_t *prog)
-{
-	struct perf_event_attr attr = {};
-	int cpu = -1, ncpus;
-	profile_t *profile;
-	char *freqstr;
-	int freq = -1;
-	int err = 0;
-
-	/*
-	 * Expected format is either profile:[n]hz where n is a number between
-	 * 1 and 1000, or profile:[c]:[n]hz where c is the CPU to profile.
-	 */
-	if (sscanf(probe->string, "profile:%dhz", &freq) != 1 &&
-	    sscanf(probe->string, "profile:%d:%dhz", &cpu, &freq) != 2)
-		return -EINVAL;
-
-	if (freq < 0 || freq > 1000)
-		return -EINVAL;
-
-	ncpus = sysconf(_SC_NPROCESSORS_ONLN);
-
-	if (cpu < -1 || cpu > ncpus)
-		return -EINVAL;
-
-	if (cpu >= 0)
-		ncpus = 1;
-
-	profile = calloc(1, sizeof(*profile));
-	if (!profile)
-		return -ENOMEM;
-
-	profile->efds = calloc(ncpus, sizeof (int));
-	if (!profile->efds) {
-		free(profile);
-		return -ENOMEM;
-	}
-	if (cpu != -1)
-		err = profile_perf_event_open(profile, cpu, freq);
-	else {
-		for (cpu = 0; cpu < ncpus; cpu++) {
-			err = profile_perf_event_open(profile, cpu, freq);
-			if (err)
-				goto out;
-		}
-		profile->num++;
-	}
-
-	if (!err)
-		err = kprobe_load(probe, prog, "kprobe:perf_swevent_hrtimer",
-				  "p", &profile->kp);
-out:
-	if (err <= 0)
-		profile_destroy(profile);
-	else
-		probe->dyn->probe.pvdr_priv = profile;
-
-	return err;
-}
-
-pvdr_t profile_pvdr = {
-        .name = "profile",
-		.desc = "Profiles with time-based intervals.",
-
-        .resolve = profile_resolve,
-
-        .setup = profile_setup,
-        .teardown = profile_teardown,
-};
-
-static int uprobe_load(node_t *probe, prog_t *prog, const char *type,
-		       kprobe_t **kpp)
-{
-	kprobe_t *kp;
-	char *func;
-
-	kp = probe_load(BPF_PROG_TYPE_KPROBE, probe, prog);
-	if (!kp)
-		return -EINVAL;
-
-	kp->type = type;
-
-	kp->ctrl = fopen("/sys/kernel/debug/tracing/uprobe_events", "a+");
-	if (!kp->ctrl) {
-		_eno("unable to open uprobe_events");
-		return -errno;
-	}
-
-	*kpp = kp;
-	func = strchr(probe->string, ':') + 1;
-	return kprobe_setattach_pattern(kp, func, 1);
-}
-
-static int uprobe_setup(node_t *probe, prog_t *prog)
-{
-	return uprobe_load(probe, prog, "p",
-			   (kprobe_t **)&probe->dyn->probe.pvdr_priv);
-}
-
-pvdr_t uprobe_pvdr = {
-	.name = "uprobe",
-	.desc = "Userland function entry provider.",
-
-	.resolve = kprobe_resolve,
-
-	.setup = uprobe_setup,
-	.teardown = kprobe_teardown,
-};
-
-static int uretprobe_setup(node_t *probe, prog_t *prog)
-{
-	return uprobe_load(probe, prog, "r",
-			   (kprobe_t **)&probe->dyn->probe.pvdr_priv);
-}
-
-pvdr_t uretprobe_pvdr = {
-	.name = "uretprobe",
-	.desc = "Userland function return provider.",
-
-	.resolve = kretprobe_resolve,
-
-	.setup = uretprobe_setup,
-	.teardown = kprobe_teardown,
-};
-
 /* Register Proiver */
 __attribute__((constructor))
 static void kprobe_pvdr_register(void)
 {
 	pvdr_register(   &kprobe_pvdr);
 	pvdr_register(&kretprobe_pvdr);
-	pvdr_register(  &profile_pvdr);
-	pvdr_register(   &uprobe_pvdr);
-	pvdr_register(&uretprobe_pvdr);
 }
